@@ -38,43 +38,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.13",
-    "@angular/common": "^7.2.13",
-    "@angular/compiler": "^7.2.13",
-    "@angular/core": "^7.2.13",
-    "@angular/elements": "^7.2.13",
-    "@angular/forms": "^7.2.13",
+    "@angular/animations": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/elements": "^8.0.0",
+    "@angular/forms": "^8.0.0",
     "@angular/http": "^7.2.13",
-    "@angular/platform-browser": "^7.2.13",
-    "@angular/platform-browser-dynamic": "^7.2.13",
-    "@angular/router": "^7.2.13",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "@ngx-loading-bar/core": "^2.2.0",
     "@ngx-loading-bar/http-client": "^2.2.0",
     "@stackblitz/sdk": "^1.2.0",
-    "@swimlane/docspa-core": "^2.0.0",
-    "@swimlane/docspa-remark-preset": "^2.0.0",
+    "@swimlane/docspa-core": "^4.0.1",
+    "@swimlane/docspa-remark-preset": "^2.0.1",
     "@swimlane/dragula": "^3.8.0",
     "@types/dragula": "^2.1.33",
     "@webcomponents/custom-elements": "^1.2.0",
     "core-js": "^2.5.4",
     "deepmerge": "^2.1.1",
-    "ngx-logger": "^3.0.5",
+    "ngx-logger": "^3.3.13",
     "path": "^0.12.7",
     "process": "^0.11.10",
     "quick-lru": "^1.1.0",
     "remark-inline-links": "^3.1.1",
-    "rxjs": "^6.4.0",
+    "rxjs": "^6.5.2",
     "smoothscroll-polyfill": "^0.4.3",
     "url-resolve": "^1.0.0",
     "vfile-reporter": "^5.0.0",
-    "zone.js": "^0.8.29"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular-devkit/build-ng-packagr": "^0.13.8",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular-devkit/build-ng-packagr": "~0.13.0",
     "@angular/cli": "^7.3.8",
-    "@angular/compiler-cli": "^7.2.13",
-    "@angular/language-service": "^7.2.13",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
     "@swimlane/prettier-config-swimlane": "^0.1.1",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
@@ -82,7 +82,7 @@
     "@types/node": "^8.9.5",
     "angular-cli-ghpages": "^0.5.2",
     "chg": "^0.3.3",
-    "codelyzer": "~4.2.1",
+    "codelyzer": "^5.1.0",
     "cpx": "^1.5.0",
     "cross-env": "^5.2.0",
     "cypress": "^3.2.0",
@@ -93,12 +93,12 @@
     "scss-bundle": "^2.5.1",
     "start-server-and-test": "^1.4.1",
     "ts-node": "~5.0.1",
-    "tsickle": "^0.34.3",
-    "tslib": "^1.7.1",
+    "tsickle": "^0.35.0",
+    "tslib": "^1.9.0",
     "tslint": "~5.9.1",
     "tslint-config-prettier": "^1.13.0",
     "tslint-config-swimlane": "^3.0.4",
-    "typescript": "~3.2.4"
+    "typescript": "~3.4.5"
   },
   "repository": {
     "type": "git",
