@@ -3,8 +3,8 @@
   "description": "Drag and Drop for Angular2 and beyond!",
   "version": "7.0.0",
   "peerDependencies": {
-    "@angular/common": "^6.0.0-rc.0 || ^6.0.0 || ^7.0.0",
-    "@angular/core": "^6.0.0-rc.0 || ^6.0.0 || ^7.0.0"
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0"
   },
   "repository": {
     "type": "git",
